# Faster `_changebasis_sim`