```julia
using BasicBSpline
using BenchmarkTools

p = 3
P1 = BSplineSpace{p}(KnotVector(1:8))
P2 = BSplineSpace{p}(KnotVector(2,3,3,4,5,6,7,8))
changebasis(P1,P2)
@benchmark BasicBSpline._changebasis_sim(P1,P2)
```

**On the current master**
```
BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  20.158 μs …  6.812 ms  ┊ GC (min … max): 0.00% … 98.70%
 Time  (median):     22.092 μs              ┊ GC (median):    0.00%
 Time  (mean ± σ):   23.798 μs ± 95.967 μs  ┊ GC (mean ± σ):  5.65% ±  1.40%

          ▂▅▆█▇▅▅▄▃                                            
  ▁▁▁▂▃▄▇████████████▇▆▆▄▅▄▄▃▃▃▃▂▂▂▂▂▂▂▂▂▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁ ▃
  20.2 μs         Histogram: frequency by time        28.3 μs <

 Memory estimate: 8.38 KiB, allocs estimate: 173.
```

**With this PR**
```
BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  14.698 μs … 85.201 μs  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     15.739 μs              ┊ GC (median):    0.00%
 Time  (mean ± σ):   16.729 μs ±  2.978 μs  ┊ GC (mean ± σ):  0.00% ± 0.00%

  ▁▆███▇▆▅▄▃▃▃▂▃▃▃▂▂▁▁                 ▁ ▁▂▁▂                 ▂
  ████████████████████▇▇▆▇▅▅▅▅▅▅▅▄▅▆▆▇▇████████▇▆▆▇▅▆▆▆▅▅▅▆▅▆ █
  14.7 μs      Histogram: log(frequency) by time      28.5 μs <

 Memory estimate: 7.53 KiB, allocs estimate: 167.
```