@hyrodium 2022-08-01
Getting faster with `bsplinebasisall`.
```julia
julia> using BasicBSpline

julia> using BenchmarkTools

julia> p = 3
3

julia> P1 = BSplineSpace{p}(KnotVector(1:8))
BSplineSpace{3, Int64}(KnotVector([1, 2, 3, 4, 5, 6, 7, 8]))

julia> P2 = BSplineSpace{p}(KnotVector(2,3,3,4,5,6,7,8))
BSplineSpace{3, Int64}(KnotVector([2, 3, 3, 4, 5, 6, 7, 8]))

julia> changebasis(P1,P2)
4×4 Matrix{Float64}:
 0.666667     0.0  -4.93432e-17   0.0
 0.333333     1.0  -2.22045e-16  -5.55112e-17
 3.70074e-17  0.0   1.0          -2.22045e-16
 0.0          0.0   0.0           1.0

julia> @benchmark BasicBSpline._changebasis_sim(P1,P2)
BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):   9.919 μs …  5.473 ms  ┊ GC (min … max): 0.00% … 99.17%
 Time  (median):     10.450 μs              ┊ GC (median):    0.00%
 Time  (mean ± σ):   11.455 μs ± 54.631 μs  ┊ GC (mean ± σ):  4.74% ±  0.99%

   ▁▇█▅▁                                                       
  ▂█████▇▅▄▃▂▂▂▃▃▃▂▂▃▂▂▂▂▂▂▂▂▂▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁ ▂
  9.92 μs         Histogram: frequency by time        16.2 μs <

 Memory estimate: 6.78 KiB, allocs estimate: 143.
```