@hyrodium 2022-08-01
:tada: :tada: :tada:
```julia
julia> @benchmark BasicBSpline._changebasis_sim(P1,P2)
BenchmarkTools.Trial: 10000 samples with 10 evaluations.
 Range (min … max):  988.800 ns … 376.767 μs  ┊ GC (min … max): 0.00% … 99.21%
 Time  (median):       1.097 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):     1.319 μs ±   6.258 μs  ┊ GC (mean ± σ):  8.18% ±  1.72%

   ▆▇██▇▆▅▅▃▂▂▂▂▂▂▂▂▂▂▃▃▃▂▂▂▂▁▂▂▁▁▁▁                 ▁          ▂
  ███████████████████████████████████▇█▇█▇█▇▇▆▇▇▆▇▇▇▇███▇▇▇▄▆▅▄ █
  989 ns        Histogram: log(frequency) by time       2.26 μs <

 Memory estimate: 1.97 KiB, allocs estimate: 25.
```