@hyrodium 2022-08-01
The test failures are related to the following issues around StaticArrays.jl.

# 0×0 matrix division
```julia
julia> using StaticArrays

julia> n = SMatrix{0,0,Float64}()
0×0 SMatrix{0, 0, Float64, 0} with indices SOneTo(0)×SOneTo(0)

julia> n \ n
0×0 SMatrix{0, 0, Float64, 0} with indices SOneTo(0)×SOneTo(0)

julia> n / n
ERROR: MethodError: one(::Type{Union{}}) is ambiguous. Candidates:
  one(::Union{Type{T}, T}) where T<:AbstractString in Base at strings/basic.jl:262
  one(::Union{Type{P}, P}) where P<:Dates.Period in Dates at /home/hyrodium/julia/julia-1.7.3/share/julia/stdlib/v1.7/Dates/src/periods.jl:54
  one(::Type{SM}) where SM<:(Union{LinearAlgebra.Adjoint{T, <:Union{StaticArray{Tuple{var"#s2"}, T, 1} where var"#s2", StaticArray{Tuple{var"#s3", var"#s4"}, T, 2} where {var"#s3", var"#s4"}}}, LinearAlgebra.Diagonal{T, <:StaticArray{Tuple{var"#s13"}, T, 1} where var"#s13"}, LinearAlgebra.Hermitian{T, <:StaticArray{Tuple{var"#s10", var"#s11"}, T, 2} where {var"#s10", var"#s11"}}, LinearAlgebra.LowerTriangular{T, <:StaticArray{Tuple{var"#s18", var"#s19"}, T, 2} where {var"#s18", var"#s19"}}, LinearAlgebra.Symmetric{T, <:StaticArray{Tuple{var"#s7", var"#s8"}, T, 2} where {var"#s7", var"#s8"}}, LinearAlgebra.Transpose{T, <:Union{StaticArray{Tuple{var"#s2"}, T, 1} where var"#s2", StaticArray{Tuple{var"#s3", var"#s4"}, T, 2} where {var"#s3", var"#s4"}}}, LinearAlgebra.UnitLowerTriangular{T, <:StaticArray{Tuple{var"#s24", var"#s25"}, T, 2} where {var"#s24", var"#s25"}}, LinearAlgebra.UnitUpperTriangular{T, <:StaticArray{Tuple{var"#s21", var"#s22"}, T, 2} where {var"#s21", var"#s22"}}, LinearAlgebra.UpperTriangular{T, <:StaticArray{Tuple{var"#s15", var"#s16"}, T, 2} where {var"#s15", var"#s16"}}, StaticArray{Tuple{var"#s1", var"#s3"}, T, 2} where {var"#s1", var"#s3"}} where T) in StaticArrays at /home/hyrodium/.julia/dev/StaticArrays/src/linalg.jl:108
  one(::Type{<:AbstractIrrational}) in Base at irrationals.jl:154
  one(::Type{T}) where T<:Number in Base at number.jl:334
Possible fix, define
  one(::Type{Union{}})
Stacktrace:
  [1] __lu(A::SMatrix{0, 0, Union{}, 0}, #unused#::Val{true})
    @ StaticArrays ~/.julia/dev/StaticArrays/src/lu.jl:112
  [2] macro expansion
    @ ~/.julia/dev/StaticArrays/src/lu.jl:85 [inlined]
  [3] _lu(A::SMatrix{0, 0, Union{}, 0}, pivot::Val{true}, check::Bool)
    @ StaticArrays ~/.julia/dev/StaticArrays/src/lu.jl:77
  [4] lu(A::SMatrix{0, 0, Union{}, 0}, pivot::Val{true}; check::Bool)
    @ StaticArrays ~/.julia/dev/StaticArrays/src/lu.jl:50
  [5] lu(A::SMatrix{0, 0, Union{}, 0}; check::Bool)
    @ StaticArrays ~/.julia/dev/StaticArrays/src/lu.jl:54
  [6] lu(A::SMatrix{0, 0, Union{}, 0})
    @ StaticArrays ~/.julia/dev/StaticArrays/src/lu.jl:54
  [7] macro expansion
    @ ~/.julia/dev/StaticArrays/src/solve.jl:55 [inlined]
  [8] _solve
    @ ~/.julia/dev/StaticArrays/src/solve.jl:45 [inlined]
  [9] \
    @ ~/.julia/dev/StaticArrays/src/solve.jl:1 [inlined]
 [10] /(A::SMatrix{0, 0, Float64, 0}, B::SMatrix{0, 0, Float64, 0})
    @ LinearAlgebra ~/julia/julia-1.7.3/share/julia/stdlib/v1.7/LinearAlgebra/src/generic.jl:1152
 [11] top-level scope
    @ REPL[4]:1
```

This is already reported in https://github.com/JuliaArrays/StaticArrays.jl/issues/1066.

# Matrix division with `SMatrix{n, n, BigFloat}`

```julia
julia> m = @SMatrix rand(BigFloat,3,3)
3×3 SMatrix{3, 3, BigFloat, 9} with indices SOneTo(3)×SOneTo(3):
 0.0533287  0.646633  0.713034
 0.203678   0.225494  0.889687
 0.560856   0.266583  0.283933

julia> m * m
3×3 SMatrix{3, 3, BigFloat, 9} with indices SOneTo(3)×SOneTo(3):
 0.534458  0.370379  0.815781
 0.555776  0.419728  0.598461
 0.243453  0.498473  0.717703

julia> m * inv(m)
3×3 SMatrix{3, 3, BigFloat, 9} with indices SOneTo(3)×SOneTo(3):
  1.0           0.0          4.31808e-78
  0.0           1.0          0.0
 -3.23856e-78  -4.31808e-78  1.0

julia> m / m
ERROR: setindex!() with non-isbitstype eltype is not supported by StaticArrays. Consider using SizedArray.
Stacktrace:
  [1] error(s::String)
    @ Base ./error.jl:33
  [2] setindex!
    @ ~/.julia/dev/StaticArrays/src/MArray.jl:39 [inlined]
  [3] macro expansion
    @ ~/.julia/dev/StaticArrays/src/indexing.jl:66 [inlined]
  [4] _setindex!_scalar
    @ ~/.julia/dev/StaticArrays/src/indexing.jl:46 [inlined]
  [5] setindex!
    @ ~/.julia/dev/StaticArrays/src/indexing.jl:42 [inlined]
  [6] macro expansion
    @ ~/.julia/dev/StaticArrays/src/indexing.jl:357 [inlined]
  [7] _setindex!
    @ ~/.julia/dev/StaticArrays/src/indexing.jl:320 [inlined]
  [8] setindex!
    @ ~/.julia/dev/StaticArrays/src/indexing.jl:274 [inlined]
  [9] _solve
    @ ~/.julia/dev/StaticArrays/src/solve.jl:36 [inlined]
 [10] \
    @ ~/.julia/dev/StaticArrays/src/solve.jl:1 [inlined]
 [11] /(A::SMatrix{3, 3, BigFloat, 9}, B::SMatrix{3, 3, BigFloat, 9})
    @ LinearAlgebra ~/julia/julia-1.7.3/share/julia/stdlib/v1.7/LinearAlgebra/src/generic.jl:1152
 [12] top-level scope
    @ REPL[19]:1
```
I'm not sure this should be fixed in StaticArrays.jl. (The error message can be updated, I guess.)