@hyrodium 2022-08-04
>I'm not sure this should be fixed in StaticArrays.jl. (The error message can be updated, I guess.)

Maybe, I'll fix this in a few weeks. (x-ref: https://github.com/JuliaArrays/StaticArrays.jl/issues/1069)